@@ -5,6 +5,7 @@
   "description": "TypeScript nightly to easily make use of latest features",
   "main": "dist_ts/index.js",
   "typings": "dist_ts/index.d.ts",
+  "type": "module",
   "bin": {
     "tsbuild": "./cli.js"
   },
@@ -29,16 +30,16 @@
   "dependencies": {
     "@pushrocks/early": "^3.0.6",
     "@pushrocks/smartcli": "^3.0.14",
-    "@pushrocks/smartfile": "^8.0.10",
+    "@pushrocks/smartfile": "^9.0.6",
     "@pushrocks/smartlog": "^2.0.44",
     "@pushrocks/smartpath": "^4.0.3",
-    "@pushrocks/smartpromise": "^3.1.6",
-    "typescript": "^4.4.2"
+    "@pushrocks/smartpromise": "^3.1.7",
+    "typescript": "4.7.0-dev.20220311"
   },
   "devDependencies": {
-    "@gitzone/tsrun": "^1.2.17",
-    "@pushrocks/tapbundle": "^3.2.14",
-    "@types/node": "^16.7.13",
+    "@gitzone/tsrun": "^1.2.19",
+    "@pushrocks/tapbundle": "^4.0.8",
+    "@types/node": "^17.0.21",
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0"
   },
